requirements.txt
google-genai
pandas
beautifulsoup4
numpy
